Move pagination and listing page rendering to templates/listing.dart
